backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
Pillow==9.5.0